Added yb-voyager cutover initiate command and remove --wait flag
